refa: move utils to Credit
